refactor(agent_backends): extract config submodule
Move the backend config storage / normalization / selection /
runtime-hash cluster out of the agent_backends facade. config.rs now
owns:

- Constants: SETTINGS_KEY, SECRET_BUCKET, BACKEND_RUNTIME_ENV_VERSION
- Wire types: BackendStatus (with `pub(super) fn new`),
  BackendListResponse, BackendSecretUpdate
- LoadedBackends and the tolerant-load pipeline
  (load_backend_configs_tolerant, load_backend_configs,
  resolve_backend_list_default), plus read_unknown_passthrough /
  read_hidden_codex_passthrough
- save_backend_configs
- Normalization helpers: normalize_backend,
  clear_manual_models_for_discovery_backend, canonical_backend_id,
  normalize_backend_id
- Model-list helpers: apply_discovered_models, backend_models_contain,
  backend_models_signature
- Selection helpers: find_backend, select_backend_for_request,
  backend_request_alias, infer_backend_for_model
- Hashing: runtime_hash, backend_kind_hash_key, harness_hash_key

Cross-submodule visibility uses pub(super); nothing leaks beyond
agent_backends except the existing wire types (already pub).

Tests covering these helpers stay in mod.rs in this commit and reach
the pub(super) functions via `use super::config::*;` — a focused
test-colocation pass will follow when the facade is trimmed in a
later commit, so each interim commit stays bisectable.

cargo fmt --check, cargo clippy -p claudette -p claudette-server -p
claudette-cli --all-targets --all-features with -Dwarnings, and
the agent_backends tests (111 passed) all pass. Test count matches
baseline; no public-surface or wire-shape change.
